# [http_ng] improve performance