--- a/gateway/src/resty/http_ng/headers.lua
+++ b/gateway/src/resty/http_ng/headers.lua
@@ -9,8 +9,8 @@ local getmetatable = getmetatable
 local type = type
 local lower = string.lower
 local upper = string.upper
-local sub = string.sub
 local ngx_re = ngx.re
+local re_gsub = ngx_re.gsub
 
 local normalize_exceptions = {
   etag = 'ETag'
@@ -27,20 +27,17 @@ local headers_mt = {
   end
 }
 
-local capitalize = function(string)
-  return upper(sub(string, 1, 1)) .. sub(string, 2)
+local capitalize = function(m)
+  return upper(m[0])
 end
 
-local regex_parts = [[[^_-]+]]
+local letter = [[\b([a-z])]]
 
-local key_parts_capitalized = function(key)
-  local parts = {}
+local capitalize_header = function(key)
+  key = re_gsub(key, '_', '-', 'jo')
+  key = re_gsub(key, letter, capitalize, 'jo')
 
-  for matches in ngx_re.gmatch(key, regex_parts, 'jo') do
-    insert(parts, capitalize(matches[0]))
-  end
-
-  return parts
+  return key
 end
 
 headers.normalize_key = function(key)
@@ -50,7 +47,7 @@ headers.normalize_key = function(key)
     return exception
   end
 
-  return concat(key_parts_capitalized(key), '-')
+  return capitalize_header(key)
 end
 
 local header_mt = {