@davidor 2018-01-19
This is great :+1: 

In the future, I think it'd be good to perform a benchmark for each optimization so we can see the impact of each of them separately.