@andrewdavidmackenzie 2018-01-31
A classic optimization trade-off result I see: averages go down, and maximums go up.